Don't refresh notesListViewModel while it's alive; Move NoteFragment to *.ui.note pkg
