+++ app/src/main/res/layout/fragment_note.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.NoteFragment">
+    tools:context=".ui.note.NoteFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
